break renderer and joint sliders into separate files
